requirements.txt
openpyxl
anthropic
pillow
